requirements.txt
gradio
llama-cpp-python
pyyaml
torch
numpy
sentence-transformers
scikit-learn
requests
python-dotenv
spacy